ci/requirements.txt
pandas
scikit-learn
joblib
azureml-core
pytest
